Extract pure fatigue analysis engine
